requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
fastapi==0.74.1
geoip2==4.5.0
h11==0.13.0
httptools==0.3.0
idna==3.3
ipwhois==1.2.0
numpy==1.22.2
pandas==1.4.1
pyasn==1.6.1
pydantic==1.9.0
python-dotenv==0.19.2
python-whois==0.7.3
PyYAML==6.0
ratelimit==2.2.1
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
termcolor==1.1.0
typing-extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
